feat: move more code to `builder`
